Fix some -Wsign-compare warnings

+++ src/actions.h
@@ -27,7 +27,7 @@
 extern int list_signatures(pesign_context *ctx);
 extern void check_signature_space(pesign_context *ctx);
 extern void allocate_signature_space(Pe *pe, ssize_t sigspace);
-extern off_t export_signature(cms_context *cms, int fd, int ascii_armor);
+extern ssize_t export_signature(cms_context *cms, int fd, int ascii_armor);
 extern void import_raw_signature(pesign_context *pctx);
 extern void remove_signature(pesign_context *ctx);
 extern void export_pubkey(pesign_context *ctx);
